agent-cdp/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
bip-utils==2.9.3
bitarray==2.9.2
cbor2==5.6.4
cdp-sdk==0.10.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
coincurve==20.0.0
crcmod==1.7
cryptography==43.0.1
cytoolz==0.12.3
distro==1.9.0
ecdsa==0.19.0
ed25519-blake2b==1.4.1
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
fastapi==0.115.5
frozenlist==1.4.1
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-core==0.3.18
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langgraph==0.2.48
langgraph-checkpoint==2.0.4
langgraph-sdk==0.1.36
langsmith==0.1.143
msgpack==1.1.0
multidict==6.1.0
numpy==1.26.4
openai==1.54.4
orjson==3.10.11
packaging==24.2
parsimonious==0.10.0
propcache==0.2.0
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rlp==4.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tenacity==9.0.0
tiktoken==0.8.0
toolz==0.12.1
tqdm==4.67.0
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
web3==7.2.0
websockets==13.0.1
yarl==1.11.1
